Refactor Entity into Vertex and Edge sharing an id field
This makes a lot of functions that can be identical for vertices and edges much easier to write
