+++ data-generation/src/main/scala/thesis/TemporalGraph.scala
@@ -3,7 +3,6 @@ package thesis
 import org.apache.spark.graphx.{EdgeRDD, EdgeTriplet, Graph, VertexId, VertexRDD}
 import org.apache.spark.rdd.RDD
 import thesis.DataTypes.{Attributes, EdgeId}
-import thesis.Entity.{EDGE, VERTEX}
 
 import java.time.Instant
 import scala.reflect.ClassTag
